application/backend/requirements.txt
fastapi==0.110.2
uvicorn==0.30.0

FlagEmbedding==1.2.10

numpy==1.26.4
scikit-learn==1.5.1
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9

requests==2.31.0
pillow==10.3.0

rank-bm25==0.2.2
openai==1.57.0
weaviate-client